Fixed Integer-Casting

+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/hibernate/AbstractDaoHibernate.java
@@ -365,10 +365,10 @@ public abstract class AbstractDaoHibernate<T, K extends Serializable> extends Hi
                 
             	final Criteria hibernateCriteria = m_criteriaConverter.convertForCount(criteria, session);
             	hibernateCriteria.setProjection(Projections.rowCount());
-                Integer retval = (Integer)hibernateCriteria.uniqueResult();
+                Long retval = (Long)hibernateCriteria.uniqueResult();
                 hibernateCriteria.setProjection(null);
                 hibernateCriteria.setResultTransformer(Criteria.ROOT_ENTITY);
-                return retval;
+                return retval.intValue();
             }
         };
         Integer retval = getHibernateTemplate().execute(callback);
